+++ python/nelsie/slidedeck.py
@@ -10,15 +10,33 @@ from .insteps import Step
 from .textstyle import TextStyle, _data_to_text_style
 
 from . import data
+
 BUILTIN_FONTS_DIR = os.path.abspath(import_resources.files(data) / "fonts")
+
 
 class Resources:
     def __init__(
-        self, *, builtin_fonts: bool = True, system_fonts: bool = False, default_code_syntaxes: bool = True, default_code_themes: bool = True
+        self,
+        *,
+        builtin_fonts: bool = True,
+        system_fonts: bool = False,
+        default_code_syntaxes: bool = True,
+        default_code_themes: bool = True,
     ):
         self._resources = nelsie_rs.Resources(system_fonts, default_code_syntaxes, default_code_themes)
         if builtin_fonts:
             self._resources.load_fonts_dir(BUILTIN_FONTS_DIR)
+            self._resources.set_generic_family("sans-serif", "DejaVu Sans")
+            self._resources.set_generic_family("monospace", "DejaVu Sans Mono")
+
+    def set_sans_serif(self, font_name):
+        self._resources.set_generic_family("sans-serif", font_name)
+
+    def set_monospace(self, font_name):
+        self._resources.set_generic_family("monospace", font_name)
+
+    def set_serif(self, font_name):
+        self._resources.set_generic_family("serif", font_name)
 
     def load_code_syntax_dir(self, path: str):
         self._resources.load_code_syntax_dir(path)
@@ -103,8 +121,6 @@ class SlideDeck:
         bg_color: str = "white",
         image_directory: str | None = None,
         resources: Resources | None = None,
-        default_font: str | None = None,
-        default_monospace_font: str | None = None,
         default_code_theme: str = "InspiredGitHub",
         default_code_language: str | None = None,
     ):
@@ -140,7 +156,7 @@ class SlideDeck:
         self.resources = resources
         self.default_code_theme = default_code_theme
         self.default_code_language = default_code_language
-        self._deck = nelsie_rs.Deck(resources._resources, default_font, default_monospace_font)
+        self._deck = nelsie_rs.Deck()
         self._slides: List[Slide] = []
 
     def set_style(self, name: str, style: TextStyle):
@@ -255,4 +271,11 @@ class SlideDeck:
         assert 0 <= compression_level <= 10
         if path:
             path = str(path)
-        return self._deck.render(self.resources._resources, verbose, output_format, compression_level, path, n_threads)
+        return self._deck.render(
+            self.resources._resources,
+            verbose,
+            output_format,
+            compression_level,
+            path,
+            n_threads,
+        )